+++ demo/src/Root.jsx
@@ -7,7 +7,7 @@ import getUserConfirmation from '../../src/utils/getUserConfirmation';
 
 const Root = ({ store, routes, nav }) => (
   <Provider store={store}>
-    <BrowserRouter getUserConfirmation={getUserConfirmation}>
+    <BrowserRouter getUserConfirmation={getUserConfirmation(store.dispatch)}>
       <App routes={routes} navConfig={nav} dispatch={store.dispatch} />
     </BrowserRouter>
   </Provider>