+++ demo/src/store.js
@@ -1,16 +1,10 @@
 import { applyMiddleware, combineReducers, createStore } from 'redux';
 import logger from 'redux-logger';
-import createSagaMiddleware from 'redux-saga';
 import preloadedState from './config/preloadedState';
 import reducers from '../../src/reducers';
-import saga from '../../src/sagas';
-
-const sagaMiddleware = createSagaMiddleware();
 
 export default createStore(
   combineReducers({ adminUi: reducers }),
   preloadedState,
-  applyMiddleware(logger, sagaMiddleware),
+  applyMiddleware(logger),
 );
-
-sagaMiddleware.run(saga);