+++ src/actions/clickDropdownTitle.js
@@ -1,8 +1,6 @@
 import { actionTypes } from '../constants';
 
-export default (navId, to, history) => ({
+export default navId => ({
   type: actionTypes.DROPDOWN_TITLE_CLICKED,
   navId,
-  to,
-  history,
 });