+++ src/actions/requesteChageRoute.js
@@ -1,0 +1,8 @@
+import { actionTypes } from '../constants';
+
+export default (navId, history, nextLocation) => ({
+  type: actionTypes.CHANGE_ROUTE_REQUESTED,
+  navId,
+  history,
+  nextLocation,
+});