+++ src/reducers/mainNav.js
@@ -2,21 +2,40 @@ import createReducer from '@triniti/app/createReducer';
 import { actionTypes } from '../constants';
 
 const {
-  CURRENT_SECTION_CHANGED,
+  CHANGE_ROUTE_CANCELLED,
+  CHANGE_ROUTE_REQUESTED,
   DROPDOWN_TITLE_CLICKED,
   MAIN_NAV_TOGGLED,
+  ROUTE_DID_CHANGED,
 } = actionTypes;
 
 export const initialState = {
   isOpen: false,
   activeSections: [],
   currentSection: '',
+  nextSection: null,
 };
 
-const onCurrentSectionChanged = (state, action) =>
+const onChangeRouteCancelled = state =>
   Object.assign({}, state, {
-    currentSection: action.navId,
+    nextSection: '',
   });
+
+const onChangeRouteRequested = (state, action) =>
+  Object.assign({}, state, {
+    nextSection: action.navId,
+  });
+
+const onRouteDidChanged = (state) => {
+  const newState = { ...state };
+  if (state.nextSection) {
+    newState.currentSection = state.nextSection;
+  }
+  newState.nextSection = '';
+  newState.isOpen = false;
+
+  return newState;
+};
 
 const onMainNavToggled = (state, action) => {
   const isOpen = typeof action.isOpen === 'undefined' ? !state.isOpen : !!action.isOpen;
@@ -31,29 +50,31 @@ const onMainNavToggled = (state, action) => {
  * @returns {*}
  */
 const onDropdownTitleClicked = (state, action) => {
-  const { activeSections } = state;
-  let updatedActiveSections = [];
+  const newState = Object.assign({}, state);
 
-  if (activeSections.includes(action.navId)) {
-    const index = activeSections.indexOf(action.navId);
-    updatedActiveSections = [
+  const { activeSections } = newState;
+  const { navId } = action;
+
+  if (activeSections.includes(navId)) {
+    const index = activeSections.indexOf(navId);
+    newState.activeSections = [
       ...activeSections.slice(0, index),
       ...activeSections.slice(index + 1),
     ];
   } else {
-    updatedActiveSections = [
+    newState.activeSections = [
       ...activeSections,
-      action.navId,
+      navId,
     ];
   }
 
-  return Object.assign({}, state, {
-    activeSections: updatedActiveSections,
-  });
+  return newState;
 };
 
 export default createReducer(initialState, {
-  [CURRENT_SECTION_CHANGED]: onCurrentSectionChanged,
+  [CHANGE_ROUTE_CANCELLED]: onChangeRouteCancelled,
+  [CHANGE_ROUTE_REQUESTED]: onChangeRouteRequested,
   [DROPDOWN_TITLE_CLICKED]: onDropdownTitleClicked,
   [MAIN_NAV_TOGGLED]: onMainNavToggled,
+  [ROUTE_DID_CHANGED]: onRouteDidChanged,
 });